Add option to send automatic beacons only once.
Automatic beacons currently have a design flaw that can cause extra
unwanted beacons to be sent out. For example, a top-level navigation
beacon is set in a click handler for an ad when the user clicks on the
ad. If the ad's visual display is programmed to change to have a button
that says "why this ad?" after the ad is clicked (the button will take
the user to an explainer page when clicked), then the automatic beacon
data will still be in place. If a user clicks back to the tab the ad is
on and clicks on "why this ad?" to perform a different top-level
navigation, the beacon will send out again as if the user clicked on the
ad normally. This creates an erroneous impression and results in
incorrect data being sent to reporting servers.

To fix the issue, this CL introduces a new parameter to the `FenceEvent`
dictionary: `once`. It defaults to false (to preserve the current
default behavior of automatic beacons). If set to true, when an
automatic beacon is sent out, the automatic beacon data will be cleared.

This CL also moves `AutomaticBeaconInfo` implementation functions from
"fenced_frame_reporter.cc" to "fenced_frame_config.cc" to reflect the
fact that its definition lives in "fenced_frame_config.h".

Change-Id: I1240557844a3c47c680a0f3d8a227e31794ab49f
Bug: 1464599